requirements.txt
scikit-learn==0.24.2
#scipy
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy
opt-einsum==3.3.0
Pillow==8.2.0
threadpoolctl==2.1.0
tensorflow-estimator==2.4.0
tensorflow-gpu== 2.2.0
#termcolor==1.1.0
#
tqdm==4.61.0
#typing-extensions==3.10.0.0
urllib3==1.26.5
#Werkzeug==2.0.1
#wrapt==1.12.1
